feat: Move controllers and update imports for better module organization
This commit moves controller files from their respective modules to a central `controllers` directory. This change promotes a more organized project structure and improves maintainability by centralizing controller logic.

The following changes were made:

- Moved `DemographicsController` to `src/controllers/demographics.controller.ts`.
- Moved `HealthController` to `src/controllers/health.controller.ts`.
- Moved `MetricsController` to `src/controllers/metrics.controller.ts`.
- Moved `HealthController` tests to `src/controllers/__tests__/health.controller.spec.ts`.
- Updated imports in `HealthModule`, `MetricsModule`, `health.integration.spec.ts`, and `health.controller.spec.ts` to reflect the new controller locations.
- Corrected import paths within `HealthController` and `MetricsController` to reference their respective services.
