+++ src/controllers/__tests__/health.controller.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing'
 import { INestApplication } from '@nestjs/common'
 import * as request from 'supertest'
 import { HealthController } from '../health.controller'
-import { HealthService, ReadinessResult } from '../health.service'
+import { HealthService, ReadinessResult } from '../../health/health.service'
 import { ConfigService } from '../../services/config.service'
 import { describe, it, expect, beforeEach, afterEach, jest } from '@jest/globals'
 import { PrismaService } from '../../services/prisma.service'
